% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_openscipkg.R
\name{create_openscipkg}
\alias{create_openscipkg}
\title{Create a new open science R package}
\usage{
create_openscipkg(package_name, path = NULL)
}
\arguments{
\item{package_name}{The name of the package to create.}

\item{path}{The directory in which to create the package. Defaults to the current working directory.}
}
\value{
A message indicating the package was created.
}
\description{
This function creates a new open science R package with a standardized structure and setup.
It also prompts the user to enter some basic information about the package and the author.
}
\examples{
\dontrun{
create_openscipkg("myPackage", path = "/myPath")
}
}
