% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\name{document_data}
\alias{document_data}
\title{Document the processed dataset}
\usage{
document_data(
  data_name,
  data_desc,
  path,
  data_format,
  data_source,
  variables = NULL,
  column_headers = FALSE
)
}
\arguments{
\item{data_name}{The name of the processed dataset.}

\item{data_desc}{The description of the processed dataset.}

\item{path}{The path to the R directory of the package.}

\item{data_format}{The format of the processed data.}

\item{data_source}{The source of the processed data.}

\item{variables}{A list of variables in the processed dataset with their descriptions.}

\item{column_headers}{Indicates whether the processed data has column headers (TRUE or FALSE).}
}
\description{
This function appends Roxygen2 documentation for the processed dataset to the existing documentation for the raw dataset.
}
\keyword{internal}
