% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{upload_data}
\alias{upload_data}
\title{Upload and rename a dataset}
\usage{
upload_data(path = NULL)
}
\arguments{
\item{path}{The directory in which to save the raw data.}
}
\value{
A message indicating the data was uploaded successfully and its new location in the 'data-raw' directory.
}
\description{
This function allows users to upload a dataset from various formats (csv, xlsx, xls, dat, sav, dta, RData, rda) and convert it to csv.
The user is also prompted to give the dataset a new name, following guidelines for good naming conventions.
}
